# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rag_System_AI - Modular Monolith",
    "description": "Scientific Journal Publication Trend Tracking System.\n\nUnified Modular Monolith RAG Pipeline covering all 12 Steps across 4 Phases:\n- Phase 1: Ingestion (Steps 1-4: Data Sources, Loading, Meaningful Chunking, Metadata Extraction)\n- Phase 2: Indexing (Steps 5-6: Embeddings, Vector Database)\n- Phase 3: Retrieval (Steps 7-9: Query Rewriting, Hybrid Search, Cross-Encoder Reranking)\n- Phase 4: Generation & Evaluation (Steps 10-12: Context Assembly, LLM Generation, Evaluation)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "Phase 1: Ingestion"
        ],
        "summary": "Upload and chunk scholarly document",
        "description": "Accepts document text, splits it into semantic chunks, and extracts initial metadata.",
        "operationId": "upload_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "Phase 1: Ingestion"
        ],
        "summary": "Get document details and chunks",
        "description": "Retrieves document metadata, processing status, and text chunks.",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique document ID",
              "examples": [
                "doc_998877"
              ],
              "title": "Document Id"
            },
            "description": "Unique document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentChunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "description": "Unique chunk identifier",
            "examples": [
              "chk_102030"
            ]
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index",
            "description": "Sequential index of the chunk",
            "examples": [
              0
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Textual body of chunk",
            "examples": [
              "Overview of scientometric methods."
            ]
          },
          "character_count": {
            "type": "integer",
            "title": "Character Count",
            "description": "Length of chunk in characters",
            "examples": [
              33
            ]
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Metadata associated with chunk"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "chunk_index",
          "content",
          "character_count"
        ],
        "title": "DocumentChunk",
        "description": "Schema representing an individual text chunk from an ingested document."
      },
      "DocumentDetailResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "examples": [
              "doc_998877"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Scientometrics_Review_2026.pdf"
            ]
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "examples": [
              "pdf"
            ]
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks",
            "examples": [
              4
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "processed"
            ]
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/DocumentChunk"
            },
            "type": "array",
            "title": "Chunks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "file_type",
          "total_chunks",
          "status"
        ],
        "title": "DocumentDetailResponse",
        "description": "Details and chunks of a stored document."
      },
      "DocumentUploadRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title or file name",
            "examples": [
              "Scientometrics_Review_2026.pdf"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Raw text content to be chunked",
            "examples": [
              "Full paper text body..."
            ]
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "File extension/type",
            "default": "txt",
            "examples": [
              "pdf"
            ]
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Metadata tags, author, year",
            "examples": [
              {
                "authors": [
                  "Nguyen Van A"
                ],
                "year": 2026
              }
            ]
          },
          "chunk_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Size",
            "description": "Target chunk character size",
            "default": 500,
            "examples": [
              500
            ]
          },
          "chunk_overlap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Overlap",
            "description": "Chunk overlap character size",
            "default": 50,
            "examples": [
              50
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "DocumentUploadRequest",
        "description": "Request schema for uploading and processing a document."
      },
      "DocumentUploadResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Generated document ID",
            "examples": [
              "doc_998877"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of document",
            "examples": [
              "Scientometrics_Review_2026.pdf"
            ]
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks",
            "description": "Count of chunks generated",
            "examples": [
              4
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "processed",
            "examples": [
              "processed"
            ]
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/DocumentChunk"
            },
            "type": "array",
            "title": "Chunks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "total_chunks"
        ],
        "title": "DocumentUploadResponse",
        "description": "Response returned after a document has been accepted and chunked."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}